requirements.txt
flask
scikit-learn
joblib
numpy
pandas
gunicorn
google-cloud-bigquery
flask-cors
